refactor: Move station editor component back to departures module as it is not shared
