+++ src/app/departures/card/favorite/departures-card-favorite.component.html
@@ -19,7 +19,7 @@
     <h2 *ngIf="monitoredStation.departures && monitoredStation.departures.length === 0 && !isUpdating">{{ 'departures.none-found' | translate}}</h2>
   </ion-card-content>
   <ion-card-content *ngIf="inEditMode">
-    <app-station-editor [monitoredStation]="monitoredStation" (submittedEvent)="onMonitoredStationEditorSubmitted()" 
-      (monitoredStationRemovedEvent)="onMonitoredStationRemoved($event)"></app-station-editor>
+    <app-departures-station-editor [monitoredStation]="monitoredStation" (submittedEvent)="onMonitoredStationEditorSubmitted()" 
+      (monitoredStationRemovedEvent)="onMonitoredStationRemoved($event)"></app-departures-station-editor>
   </ion-card-content>
 </ion-card>
