+++ src/app/departures/station-editor/departures-station-editor.component.spec.ts
@@ -1,19 +1,19 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
-import { StationEditorComponent } from './station-editor.component';
+import { DeparturesStationEditorComponent } from './departures-station-editor.component';
 
-describe('StationEditorComponent', () => {
-  let component: StationEditorComponent;
-  let fixture: ComponentFixture<StationEditorComponent>;
+describe('DeparturesStationEditorComponent', () => {
+  let component: DeparturesStationEditorComponent;
+  let fixture: ComponentFixture<DeparturesStationEditorComponent>;
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ StationEditorComponent ],
+      declarations: [ DeparturesStationEditorComponent ],
       imports: [IonicModule.forRoot()]
     }).compileComponents();
 
-    fixture = TestBed.createComponent(StationEditorComponent);
+    fixture = TestBed.createComponent(DeparturesStationEditorComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   }));
